# specs/001-cms-automation/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: CMS Automation API
  description: |
    API for AI-powered content management system automation.
    Enables automated article generation, tagging, scheduling, and publishing workflows.
  version: 1.0.0
  contact:
    name: CMS Automation Team
  license:
    name: MIT
servers:
  - url: https://api.cms-automation.example.com/v1
    description: Production server
  - url: https://staging-api.cms-automation.example.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /topics:
    post:
      tags: [Topics]
      summary: Submit article topic request
      description: |
        Submit a new topic for AI article generation. Supports single topics
        or batch submissions. Triggers async article generation workflow.
      operationId: submitTopicRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopicRequestCreate'
      responses:
        '202':
          description: Topic request accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicRequestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'

    get:
      tags: [Topics]
      summary: List topic requests
      description: Retrieve paginated list of topic requests with filtering
      operationId: listTopicRequests
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          description: Filter by request status
          schema:
            type: string
            enum: [pending, processing, completed, failed, cancelled]
        - name: priority
          in: query
          description: Filter by priority
          schema:
            type: string
            enum: [urgent, normal, low]
      responses:
        '200':
          description: Topic requests retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicRequestList'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /topics/{requestId}:
    get:
      tags: [Topics]
      summary: Get topic request details
      description: Retrieve detailed information about a specific topic request
      operationId: getTopicRequest
      parameters:
        - $ref: '#/components/parameters/RequestIdParam'
      responses:
        '200':
          description: Topic request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicRequestDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Topics]
      summary: Cancel topic request
      description: Cancel a pending or processing topic request
      operationId: cancelTopicRequest
      parameters:
        - $ref: '#/components/parameters/RequestIdParam'
      responses:
        '204':
          description: Topic request cancelled
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # ARTICLES
  # ============================================================================
  /articles:
    get:
      tags: [Articles]
      summary: List articles
      description: Retrieve paginated list of articles with filtering and search
      operationId: listArticles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          description: Filter by article status
          schema:
            type: string
            enum: [draft, in-review, scheduled, published, failed]
        - name: tags
          in: query
          description: Filter by tag IDs (comma-separated)
          schema:
            type: string
        - name: search
          in: query
          description: Search in title and body
          schema:
            type: string
        - name: author_id
          in: query
          description: Filter by author user ID
          schema:
            type: integer
      responses:
        '200':
          description: Articles retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleList'
  /articles/{articleId}:
    get:
      tags: [Articles]
      summary: Get article details
      description: Retrieve complete article with metadata and workflow state
      operationId: getArticle
      parameters:
        - $ref: '#/components/parameters/ArticleIdParam'
      responses:
        '200':
          description: Article details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Articles]
      summary: Update article
      description: |
        Update article content, metadata, or tags. Supports both manual edits
        and AI re-generation requests.
      operationId: updateArticle
      parameters:
        - $ref: '#/components/parameters/ArticleIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleUpdate'
      responses:
        '200':
          description: Article updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    delete:
      tags: [Articles]
      summary: Delete article
      description: Soft delete article (moves to deleted status)
      operationId: deleteArticle
      parameters:
        - $ref: '#/components/parameters/ArticleIdParam'
      responses:
        '204':
          description: Article deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from CMS authentication system

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ArticleIdParam:
      name: articleId
      in: path
      required: true
      description: Article unique identifier
      schema:
        type: integer
    RequestIdParam:
      name: requestId
      in: path
      required: true
      description: Topic request unique identifier
      schema:
        type: integer
...
  schemas:
...
    TopicRequestCreate:
      type: object
      required:
        - topic_description
      properties:
        topic_description:
          type: string
          minLength: 10
          maxLength: 5000
          description: Topic description or outline
          example: "Write a comprehensive guide on implementing OAuth2 authentication in Python web applications"
        outline:
          type: string
          nullable: true
          description: Optional structured outline (Markdown or JSON)
          example: |
            # Introduction
            # OAuth2 Fundamentals
            # Implementation Steps
            # Security Best Practices
        style_tone:
          type: string
          enum: [professional, casual, technical, conversational, academic]
          default: professional
          description: Requested writing style
        target_word_count:
          type: integer
          minimum: 100
          maximum: 10000
          default: 1000
          description: Desired article length in words
        priority:
          type: string
          enum: [urgent, normal, low]
          default: normal
          description: Processing priority
    TopicRequestResponse:
      type: object
      properties:
        request_id:
          type: integer
          example: 42
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
          example: "pending"
        estimated_completion:
          type: string
          format: date-time
          description: Estimated completion time (generation < 5 minutes)
          example: "2025-10-25T14:35:00Z"
        message:
          type: string
          example: "Topic request accepted for processing"
    TopicRequestDetail:
      allOf:
        - $ref: '#/components/schemas/TopicRequestCreate'
        - type: object
          properties:
            id:
              type: integer
              example: 42
            status:
              type: string
              enum: [pending, processing, completed, failed, cancelled]
            submitted_by:
              type: integer
              description: User ID who submitted request
            submitted_at:
              type: string
              format: date-time
            article_id:
              type: integer
              nullable: true
              description: Generated article ID (if completed)
            error_message:
              type: string
              nullable: true
              description: Error details if generation failed
    TopicRequestList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TopicRequestDetail'
        total:
          type: integer
          example: 150
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 8
    Article:
      type: object
      properties:
        id:
          type: integer
          example: 123
        title:
          type: string
          minLength: 10
          maxLength: 500
          example: "Implementing OAuth2 Authentication in Python"
        body:
          type: string
          description: Article content (Markdown or HTML)
        status:
          type: string
          enum: [draft, in-review, scheduled, published, failed]
        author_id:
          type: integer
        created_at:
          type: string
          format: date-time
        published_at:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          description: CMS-specific metadata
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
    ArticleDetail:
      allOf:
        - $ref: '#/components/schemas/Article'
        - type: object
          properties:
            workflow_state:
              $ref: '#/components/schemas/WorkflowState'
            schedule:
              $ref: '#/components/schemas/Schedule'
              nullable: true
            version:
              type: integer
              description: Article version number
            modification_history:
              type: array
              items:
                type: object
                properties:
                  version:
                    type: integer
                  timestamp:
                    type: string
                    format: date-time
                  change_type:
                    type: string
                    enum: [manual_edit, ai_regeneration]
                  user_id:
                    type: integer
    ArticleList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Article'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
    ArticleUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 10
          maxLength: 500
        body:
          type: string
        metadata:
          type: object
          additionalProperties: true
        tag_ids:
          type: array
          items:
            type: integer
          description: Replace existing tags
...
    Tag:
      type: object
      properties:
        id:
          type: integer
          example: 7
        name:
          type: string
          example: "OAuth2"
        slug:
          type: string
          example: "oauth2"
        category:
          type: string
          nullable: true
          example: "Authentication"
        usage_count:
          type: integer
          example: 42
        source:
          type: string
          enum: [automated, manual]
        cms_tag_id:
          type: string
          nullable: true
...
    Schedule:
      type: object
      properties:
        id:
          type: integer
        article_id:
          type: integer
        scheduled_time:
          type: string
          format: date-time
          description: Target publication time (UTC)
        creator_id:
          type: integer
        status:
          type: string
          enum: [pending, published, failed, cancelled]
        retry_count:
          type: integer
        max_retries:
          type: integer
          default: 3
        created_at:
          type: string
          format: date-time
        executed_at:
          type: string
          format: date-time
          nullable: true
        error_message:
          type: string
          nullable: true
...
    WorkflowState:
      type: object
      properties:
        id:
          type: integer
        article_id:
          type: integer
        current_status:
          type: string
          enum: [pending_review, approved, rejected, revision_requested]
        assigned_reviewers:
          type: array
          items:
            type: integer
          description: User IDs of assigned reviewers
        approval_history:
          type: array
          items:
            type: object
            properties:
              reviewer_id:
                type: integer
              action:
                type: string
                enum: [approved, rejected]
              comment:
                type: string
              timestamp:
                type: string
                format: date-time
        modification_requests:
          type: array
          items:
            type: object
            properties:
              requester_id:
                type: integer
              request_type:
                type: string
                enum: [ai_regeneration, manual_edit]
              feedback:
                type: string
              status:
                type: string
                enum: [pending, completed, cancelled]
              created_at:
                type: string
                format: date-time
        version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid request parameters"
            details:
              type: object
              additionalProperties: true
            request_id:
              type: string
              format: uuid
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict (e.g., duplicate, version mismatch)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Requests remaining in window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Time when limit resets (Unix timestamp)